update eth backend
